requirements.txt
fastapi==0.104.1
pymongo==4.6.0
uvicorn==0.24.0.post1
paho-mqtt==1.6.1